feat(cache): array handler
handler array handler

+++ src/main/java/org/fz/nettyx/serializer/struct/basic/c/signed/Cchar.java
@@ -31,7 +31,7 @@ public class Cchar extends CBasic<Byte> {
      * @param value the length
      */
     public Cchar(Object value) {
-        super(value, 1);
+        super(((Integer) value).byteValue(), 1);
     }
 
     /**
